requirements.txt
# Core ML/Data Science
torch>=2.0.0
torch-geometric>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Training utilities
tensorboard>=2.13.0

# Data handling
python-dateutil>=2.8.2

# Visualization & Latent Space Mapping
plotly>=5.15.0
umap-learn>=0.5.3

# Development (optional)
# jupyter>=1.0.0
matplotlib>=3.7.0
streamlit>=1.24.0
# seaborn>=0.12.0
